disk_limiter: return currLimit(), rather than limit
`getDiskLimitInfo` was returning the max limit for the journal, rather
than the limit calculated based on available space.  Oops!

Turns out this is a float64, so all those types needed to change
sadly.  When reporting to the user, it formats the float with no
decimals.

+++ libkbfs/tlf_journal.go
@@ -1737,8 +1737,8 @@ type ErrDiskLimitTimeout struct {
 	availableFiles int64
 	usageBytes     int64
 	usageFiles     int64
-	limitBytes     int64
-	limitFiles     int64
+	limitBytes     float64
+	limitFiles     float64
 	err            error
 	reportable     bool
 }
